+++ src/types/index.ts
@@ -1,2 +1,2 @@
 export * from './internal';
-export * from './user-land';
+export * from './public';
